gaia: rework guts of the binaries
...

minor compile fix

org
